added getTimeOffset
