+++ superset-frontend/packages/superset-ui-core/package.json
@@ -51,6 +51,7 @@
     "jed": "^1.1.1",
     "lodash": "^4.17.11",
     "math-expression-evaluator": "^1.3.8",
+    "moment": "^2.30.1",
     "pretty-ms": "^7.0.0",
     "react-error-boundary": "^1.2.5",
     "react-markdown": "^8.0.3",
